backend/requirements.txt
# Minimal requirements - Core functionality only
# Install in stages to avoid dependency resolution issues

# Stage 1: Core LLM clients
openai>=1.50.0
anthropic>=0.40.0
google-generativeai>=0.8.0

# Stage 2: Vector DB and embeddings  
chromadb>=0.5.0
sentence-transformers>=3.0.0

# Stage 3: Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
python-multipart

# Stage 4: Utilities
python-dotenv
requests
beautifulsoup4
lxml
rich
typer
loguru
numpy
Pillow>=10.4.0

# Stage 5: OCR (optional - install separately if needed)
# easyocr
# pytesseract
# opencv-python
